Setup first code round
Setup of the first code to get started with the timer project. The maven command
'package' will build a war-file. This creates the simple timer application.
Every client that connects to the server will share the same count down time.
The web page show the time left of the current count down.