Multiple countdown timers
Users can add their countdown time to the list of count down timers. When the
current count down time is finished, the first value of the timer list is used
as the new count down time. The web page lists and refreshes the current timer
and the list of count down times that all the users put in.