Introduce stop and reset method
The stop() method was also used to reset the timer. This leads to an unwanted
side effect of resetting the timer after one of the anneTimes runs out. The
extra method is introduced to seperate the 'reset and stop' from the 'stop'
functionality.

Perhaps introduce a reset() method and have stopAndReset() call stop() and
reset() seperately.